requirements.txt
flask==3.0.3
flask_caching==2.3.0
spotipy==2.24.0
python_dotenv==1.0.1